+++ api/CHANGELOG.md
@@ -44,6 +44,7 @@
 - Added guards to the patcher logger to prevent massive patch log exports
 - Made the expandable/collapsable FAB buttons on the `Apps` and `Patch Bundles` tabs states persist
 - Made saved patched app entries in the `Apps` tab not overwrite each other unless the app has the same package name and was patched with the same patch bundle
+- Added the ability to set a image of choice as the app background
 
 
 # Bug fixes